Introduce DFAEdgeCache class.
